introduce HasBlock utility
